Remove support for Climate and Re-enable Networking modules
- Removes support for the retired Climate Module
- Re-enables the Networking Module to allow calls to network resources on the ISY.
- General clean-up and consolidation
