+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
     rev: v1.16.0
   - hooks:
       - additional_dependencies: [flake8-docstrings==1.5.0, pydocstyle==5.0.2]
-        files: ^(pyisy)/((?!climate).)+\.py$
+        files: ^(pyisy)/.+\.py$
         id: flake8
     repo: https://gitlab.com/pycqa/flake8
     rev: 3.7.9
